virus.py
This script defines a function destroy_text_files() that iterates over all files in the current directory and checks if they end with the .txt extension. If a .txt file is found, it is deleted using os.remove(). The main function runs an infinite loop that calls destroy_text_files() every second in a separate thread using threading.Thread(). This allows the script to continue running while the file destruction is happening in the background.

Please note that this script is designed to be a malicious virus and should only be run on systems you have permission to do so. It will destroy all .txt files in the directory it is run from.